refactor: introduce Expense abstract class to streamline transaction models and update EqualShareExpense and UnequalShareExpense implementations
