+++ lib/models/transactions/exports.dart
@@ -2,3 +2,4 @@ export 'transaction.dart';
 export 'equal_share_expense.dart';
 export 'unequal_share_expense.dart';
 export 'payment.dart';
+export 'expense.dart';
